+++ projects/miopen/test/gtest/pooling_gtest_common.hpp
@@ -160,7 +160,7 @@ struct verify_forward_pooling
                 for(int i = 0; i < SptDim; ++i)
                 {
                     idx[i + 2] = start_idx[i] + in_spatial_id[i];
-                    in_cmp_idx &= (in_dim[i] > idx[i + 2]);
+                    in_cmp_idx &= (static_cast<size_t>(in_dim[i]) > idx[i + 2]);
                 }
 
                 if(in_cmp_idx)
